Consistent route naming and broken out instance routes
